[styles] Make StyleRules backwards compatible (#16200)
* Refactor needed types to @material-ui/types

* Revert to old types if there are no props

* Add typescript version of DynamicInlineStyle

* docs:typescript:formatted
